moved processor registry to processing
