+++ src/piddiplatsch/processing/registry.py
